Stopped on the problem of getting a list of localizations Despite running a dynamic object, Laravel caches state.

+++ src/Services/Processors/Uninstall.php
@@ -26,7 +26,9 @@ final class Uninstall extends Processor
     {
         $this->log('Deleting a file: ' . $this->target_path);
 
-        File::delete($this->target_path);
+        if (File::exists($this->target_path)) {
+            File::delete($this->target_path);
+        }
     }
 
     protected function deleteDirectory(): void
